requirements.txt
websocket-client
pystray
pillow
ttkbootstrap
pyperclip
pyinstaller
